move count_wire_type_sizes() to utils
